+++ PersonalFinance-API/Domain/Entities/T_Transaction.cs
@@ -27,7 +27,7 @@ namespace PersonalFinance.Domain.Entities
         /// <summary>
         /// Transaction amount
         /// </summary>
-        public double Amount { get; set; } = 0;
+        public decimal Amount { get; set; } = 0;
         /// <summary>
         /// Date and time the Transaction occurred
         /// </summary>
